refactor typing
