+++ packages/cache-persist/examples/react-example/package.json
@@ -12,7 +12,7 @@
     "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
     "styled-components": "4.2.0",
-    "@wora/cache-persist": "file:../../wora-cache-persist-2.0.0-a15.tgz"
+    "@wora/cache-persist": "file:../../wora-cache-persist-2.0.0-a16.tgz"
   },
   "scripts": {
     "start": "react-scripts-ts start",
