+++ packages/cache-persist/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wora/cache-persist",
-  "version": "2.0.0-a15",
+  "version": "2.0.0-a16",
   "keywords": [
     "cache",
     "react-native",
