+++ packages/cache-persist/src/index.ts
@@ -1,4 +1,6 @@
 import Cache from "./Cache";
 export * from "./Cache";
 
+export * from "./CacheTypes";
+
 export default Cache;
