+++ packages/cache-persist/src/layers/jsonSerialize.ts
@@ -1,4 +1,4 @@
-import { Layer } from '../StorageProxy';
+import { Layer } from '../CacheTypes';
 
 const jsonSerialize: Layer<any> = {
     set: (key: string, value: any) => { return { key, value: JSON.stringify(value)} },
